Moved Course Entity to Domain Layer, if needed a Mapper and ViewModel will be add
