+++ app/public/domain/course.service.mock.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { CourseService } from "./course.service";
-import { Course } from '../presentation/common/courses/view-model/course.view-model';
+import { Course } from './entities/course.entity';
 import { COURSES } from './courses-mock';
 
 @Injectable()
