+++ app/public/domain/entities/course.entity.ts
