+++ app/public/presentation/common/courses/course.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { Course } from './view-model/course.view-model';
+import { Course } from '../../../domain/entities/course.entity';
 
 @Component({
   selector: 'course',
